+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 
 - `cargo_clippy_args`: a rule for `cargo clippy` arguments that must be
   separated by `--`.
+- Parallel search for fixes via `rayon`.
 
 ### Changed